Flyttet grupperingsId fra eventer til nøkkel.
