+++ src/main/avro/done.avsc
@@ -7,10 +7,6 @@
             "name": "tidspunkt",
             "type": "long",
             "logicalType": "timestamp-millis"
-        },
-        {
-            "name": "grupperingsId",
-            "type": "string"
         }
     ]
 }
