+++ src/main/avro/oppgave.avsc
@@ -9,10 +9,6 @@
             "logicalType": "timestamp-millis"
         },
         {
-            "name": "grupperingsId",
-            "type": "string"
-        },
-        {
             "name": "tekst",
             "type": "string"
         },
