+++ src/main/java/no/nav/brukernotifikasjon/schemas/builders/BeskjedBuilder.java
@@ -15,7 +15,6 @@ public class BeskjedBuilder {
 
     private LocalDateTime tidspunkt;
     private LocalDateTime synligFremTil;
-    private String grupperingsId;
     private String tekst;
     private URL link;
     private Integer sikkerhetsnivaa;
@@ -29,11 +28,6 @@ public class BeskjedBuilder {
 
     public BeskjedBuilder withSynligFremTil(LocalDateTime synligFremTil) {
         this.synligFremTil = synligFremTil;
-        return this;
-    }
-
-    public BeskjedBuilder withGrupperingsId(String grupperingsId) {
-        this.grupperingsId = grupperingsId;
         return this;
     }
 
@@ -68,7 +62,6 @@ public class BeskjedBuilder {
         return new Beskjed(
                 ValidationUtil.localDateTimeToUtcTimestamp(tidspunkt, "tidspunkt", ValidationUtil.IS_REQUIRED_TIDSPUNKT),
                 ValidationUtil.localDateTimeToUtcTimestamp(synligFremTil, "synligFremTil", ValidationUtil.IS_REQUIRED_SYNLIGFREMTIL),
-                ValidationUtil.validateNonNullFieldMaxLength(grupperingsId, "grupperingsId", ValidationUtil.MAX_LENGTH_GRUPPERINGSID),
                 ValidationUtil.validateNonNullFieldMaxLength(tekst, "tekst", ValidationUtil.MAX_LENGTH_TEXT_BESKJED),
                 ValidationUtil.validateLinkAndConvertToString(link, "link", ValidationUtil.MAX_LENGTH_LINK, ValidationUtil.isLinkRequired(Eventtype.BESKJED)),
                 ValidationUtil.validateSikkerhetsnivaa(sikkerhetsnivaa),
