+++ src/main/java/no/nav/brukernotifikasjon/schemas/builders/DoneBuilder.java
@@ -9,22 +9,15 @@ import java.time.LocalDateTime;
 public class DoneBuilder {
 
     private LocalDateTime tidspunkt;
-    private String grupperingsId;
 
     public DoneBuilder withTidspunkt(LocalDateTime tidspunkt) {
         this.tidspunkt = tidspunkt;
         return this;
     }
 
-    public DoneBuilder withGrupperingsId(String grupperingsId) {
-        this.grupperingsId = grupperingsId;
-        return this;
-    }
-
     public Done build() {
         return new Done(
-                ValidationUtil.localDateTimeToUtcTimestamp(tidspunkt, "tidspunkt", ValidationUtil.IS_REQUIRED_TIDSPUNKT),
-                ValidationUtil.validateNonNullFieldMaxLength(grupperingsId, "grupperingsId", ValidationUtil.MAX_LENGTH_GRUPPERINGSID)
+                ValidationUtil.localDateTimeToUtcTimestamp(tidspunkt, "tidspunkt", ValidationUtil.IS_REQUIRED_TIDSPUNKT)
         );
     }
 }
