+++ src/main/java/no/nav/brukernotifikasjon/schemas/builders/OppgaveBuilder.java
@@ -14,7 +14,6 @@ import java.util.List;
 public class OppgaveBuilder {
 
     private LocalDateTime tidspunkt;
-    private String grupperingsId;
     private String tekst;
     private URL link;
     private Integer sikkerhetsnivaa;
@@ -23,11 +22,6 @@ public class OppgaveBuilder {
 
     public OppgaveBuilder withTidspunkt(LocalDateTime tidspunkt) {
         this.tidspunkt = tidspunkt;
-        return this;
-    }
-
-    public OppgaveBuilder withGrupperingsId(String grupperingsId) {
-        this.grupperingsId = grupperingsId;
         return this;
     }
 
@@ -61,7 +55,6 @@ public class OppgaveBuilder {
     public Oppgave build() {
         return new Oppgave(
                 ValidationUtil.localDateTimeToUtcTimestamp(tidspunkt, "tidspunkt", ValidationUtil.IS_REQUIRED_TIDSPUNKT),
-                ValidationUtil.validateNonNullFieldMaxLength(grupperingsId, "grupperingsId", ValidationUtil.MAX_LENGTH_GRUPPERINGSID),
                 ValidationUtil.validateNonNullFieldMaxLength(tekst, "tekst", ValidationUtil.MAX_LENGTH_TEXT_OPPGAVE),
                 ValidationUtil.validateLinkAndConvertToString(link, "link", ValidationUtil.MAX_LENGTH_LINK, ValidationUtil.isLinkRequired(Eventtype.OPPGAVE)),
                 ValidationUtil.validateSikkerhetsnivaa(sikkerhetsnivaa),
