+++ src/main/java/no/nav/brukernotifikasjon/schemas/builders/StatusoppdateringBuilder.java
@@ -12,7 +12,6 @@ import java.time.LocalDateTime;
 public class StatusoppdateringBuilder {
 
     private LocalDateTime tidspunkt;
-    private String grupperingsId;
     private URL link;
     private Integer sikkerhetsnivaa;
     private StatusGlobal statusGlobal;
@@ -21,11 +20,6 @@ public class StatusoppdateringBuilder {
 
     public StatusoppdateringBuilder withTidspunkt(LocalDateTime tidspunkt) {
         this.tidspunkt = tidspunkt;
-        return this;
-    }
-
-    public StatusoppdateringBuilder withGrupperingsId(String grupperingsId) {
-        this.grupperingsId = grupperingsId;
         return this;
     }
 
@@ -57,7 +51,6 @@ public class StatusoppdateringBuilder {
     public Statusoppdatering build() {
         return new Statusoppdatering(
                 ValidationUtil.localDateTimeToUtcTimestamp(tidspunkt, "tidspunkt", ValidationUtil.IS_REQUIRED_TIDSPUNKT),
-                ValidationUtil.validateNonNullFieldMaxLength(grupperingsId, "grupperingsId", ValidationUtil.MAX_LENGTH_GRUPPERINGSID),
                 ValidationUtil.validateLinkAndConvertToString(link, "link", ValidationUtil.MAX_LENGTH_LINK, ValidationUtil.isLinkRequired(Eventtype.STATUSOPPDATERING)),
                 ValidationUtil.validateSikkerhetsnivaa(sikkerhetsnivaa),
                 ValidationUtil.validateStatusGlobal(statusGlobal),
