+++ src/test/java/no/nav/brukernotifikasjon/schemas/builders/BeskjedAvroTest.java
@@ -36,7 +36,6 @@ public class BeskjedAvroTest {
     private Beskjed getBeskjedWithDefaultValues() {
         return Beskjed.newBuilder()
                 .setTidspunkt(LocalDateTime.now().toEpochSecond(ZoneOffset.UTC))
-                .setGrupperingsId("3456789123456")
                 .setTekst("Dette er informasjon du må lese")
                 .setLink("https://gyldig.url")
                 .build();
